requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
passlib
requests
pydantic
bs4
python-multipart
PyJWT
pyreadline3
pytest-cov
pytest